requirements.txt
###### Requirements without Version Specifiers ######
gym-retro
numpy
opencv-python
neat-python